Change MapView to extend CollectionView instead of ContainerView
app.js:
The general principle behind the design is the same, but instead of
extending ContainerView, the MapView now extends CollectionView.
CollectionView is more appropriate for displaying an array of data.  It
automatically creates, destroys, and keeps track of its child view
objects(in this case, MarkerView) as the underlying model changes.
Specifically, this gets rid of the pushObject calls that were being
made in the containerView when a new event was added.  Now, a
MarkerView is automatically created and its didInsertElement function
is called. We override the function to add a marker to the map.
Also, removing markers is simplified.  You no longer need a boolean,
'removed', on the EventController.  When the element is removed from
the controller, the MarkerView's willDestroyElement function is called
and we set the marker's map property to null, so it is removed from the
map.
By taking advantage of more powerful classes provided by Ember, I think
the changes make the code a little simpler and more robust.

index.html:
1) created an events controller, instead of using the more ambiguous,
Index controller
2) made the map_canvas its own div instead of declaring it inside the
ember view
3) changed some names

style.css:
1) set the style of the map canvas as we are no longer doing this in
the app.js file
2) remove trailing spaces (my editor made me)